requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
urllib3==2.2.2
snowflake-connector-python[pandas]==3.11.0
sqlalchemy
pandas==2.2.2
cryptography
python-dotenv